removing inputs as separate components
