+++ src/atoms/Fields/SelectField.tsx
@@ -1,10 +1,32 @@
 import * as React from "react"
+import styled from "styled-components"
 
 import { FieldSize } from "../../utils/constants"
 import { FieldProps } from "../../utils/types"
+import { ChevronDownIcon } from "../Icons"
 
-import SelectInput from "./Inputs/SelectInput"
+import { ContainerStyle } from "./shared"
 import Field from "./Field"
+
+const Styled = {
+  Select: styled.select`
+    appearance: none;
+    background-color: white;
+
+    &:not(:disabled):hover {
+      cursor: pointer;
+    }
+
+    ${ContainerStyle};
+  `,
+  ChevronDownIcon: styled(ChevronDownIcon)`
+    position: absolute;
+    top: 13px;
+    right: 6px;
+    color: #aaaaaa;
+  `,
+  Option: styled.option``,
+}
 
 export const SelectField = <Value extends string>(props: FieldProps<Value>) => {
   const {
@@ -19,6 +41,12 @@ export const SelectField = <Value extends string>(props: FieldProps<Value>) => {
     onAction,
   } = props
 
+  const _onChange = React.useCallback(
+    (e: React.ChangeEvent<HTMLSelectElement>) =>
+      onChange?.(e.target.value as Value),
+    [onChange],
+  )
+
   const hasError = !!error
   return (
     <Field
@@ -30,7 +58,12 @@ export const SelectField = <Value extends string>(props: FieldProps<Value>) => {
       onAction={onAction}
       error={error}
     >
-      <SelectInput value={value} onChange={onChange} hasError={hasError} />
+      <Styled.Select value={value} onChange={_onChange} $hasError={hasError}>
+        <Styled.Option value="Test 1">Test 1</Styled.Option>
+        <Styled.Option value="Test 2">Test 2</Styled.Option>
+        <Styled.Option value="Test 3">Test 3</Styled.Option>
+      </Styled.Select>
+      <Styled.ChevronDownIcon />
     </Field>
   )
 }
